relocate some query utils to runtime core
