+++ runtime/protocol/http/common/src/aws/smithy/kotlin/runtime/http/util/Text.kt
@@ -5,36 +5,8 @@
 package aws.smithy.kotlin.runtime.http.util
 
 import aws.smithy.kotlin.runtime.InternalApi
-import aws.smithy.kotlin.runtime.net.QueryParameters
-import aws.smithy.kotlin.runtime.net.QueryParametersBuilder
 import aws.smithy.kotlin.runtime.util.text.VALID_PCHAR_DELIMS
 import aws.smithy.kotlin.runtime.util.text.encodeUrlPath
-import aws.smithy.kotlin.runtime.util.text.splitAsQueryString
-
-/**
- * Parse a set of [QueryParameters] out of a full URI. If the URI does not contain a `?` (or contains nothing after the
- * `?`) then the result is null.
- */
-@InternalApi
-public fun CharSequence.fullUriToQueryParameters(): QueryParameters? {
-    val idx = indexOf("?")
-    if (idx < 0 || idx + 1 >= length) return null
-
-    val fragmentIdx = indexOf("#", startIndex = idx)
-    val rawQueryString = if (fragmentIdx > 0) substring(idx + 1, fragmentIdx) else substring(idx + 1)
-    return rawQueryString.splitAsQueryParameters()
-}
-
-/**
- * Split a (decoded) query string "foo=baz&bar=quux" into it's component parts
- */
-public fun String.splitAsQueryParameters(): QueryParameters {
-    val builder = QueryParametersBuilder()
-    splitAsQueryString().entries.forEach { entry ->
-        builder.appendAll(entry.key, entry.value)
-    }
-    return builder.build()
-}
 
 // RFC-3986 §3.3 allows sub-delims (defined in section2.2) to be in the path component.
 // This includes both colon ':' and comma ',' characters.
